+++ freecad/Mesh_Remodel/init_gui.py
@@ -1,236 +1,6 @@
 # SPDX-License-Identifier: LGPL-2.1-or-later
-# SPDX-FileCopyrightText: 2019 Mark Ganson <mwganson@gmail.com>
 # SPDX-FileNotice: Part of the Mesh Remodel addon.
 
-################################################################################
-#                                                                              #
-#   This addon is free software: you can redistribute it and/or modify it      #
-#   under the terms of the GNU Lesser General Public License as published      #
-#   by the Free Software Foundation, either version 2.1 of the License,        #
-#   or (at your option) any later version.                                     #
-#                                                                              #
-#   This addon is distributed in the hope that it will be useful,              #
-#   but WITHOUT ANY WARRANTY; without even the implied warranty                #
-#   of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.                    #
-#   See the GNU Lesser General Public License for more details.                #
-#                                                                              #
-#   You should have received a copy of the GNU Lesser General Public           #
-#   License along with this addon. If not, see https://www.gnu.org/licenses    #
-#                                                                              #
-################################################################################
-
-from FreeCAD import Gui , Console , ParamGet
-from .Misc import asIcon
-
-#def myFunc(string):
-#    print (string)
-#    global act
-#    act.setVisible(True)
-
-#mw=Gui.getMainWindow()
-#bar=mw.menuBar()
-#act=bar.addAction("MyCmd")
-#mw.workbenchActivated.connect(myFunc)
-
-####################################################################################
-# Initialize the workbench
-class MeshRemodelWorkbench(Gui.Workbench):
-
-    MenuText = "Mesh Remodel"
-    ToolTip = "MeshRemodel workbench"
-    Icon = asIcon('CreatePointsObject')
-
-    def __init__(self):
-        pass
-
-    def Initialize(self):
-        "This function is executed when FreeCAD starts"
-        self.list = [
-                    "MeshRemodelGroupCommandPointsObjects",
-                  #  "MeshRemodelCreatePointsObject",
-                  #  "MeshRemodelCreateWireFrameObject",
-                    "MeshRemodelMeshBoundaryWires",
-                    "MeshRemodelCreateCrossSectionsObject",
-                    "MeshRemodelMovePoint",
-                    "MeshRemodelRemovePoint",
-                    "MeshRemodelAddOrRemoveFacet",
-                    "MeshRemodelExpandedMesh",
-                    "MeshRemodelMeshSimpleCopy",
-                    "MeshRemodelOffsetMesh",
-                    "MeshRemodelDuplicateSelectedFacets",
-                  #  "MeshRemodelCreateCoplanarPointsObject",
-                    "MeshRemodelAddSelectionObserver",
-                    "MeshRemodelPartSolid",
-                    "MeshRemodelSubObjectLoft",
-                    "MeshRemodelCreateGridSurface",
-                    "MeshRemodelCreatePointObject",
-                    "MeshRemodelCreateLine",
-                    "MeshRemodelCreatePolygon",
-                    "MeshRemodelCreateBSpline",
-                    "MeshRemodelCreatePlane",
-                    "MeshRemodelCreateCircle",
-                    "MeshRemodelCreateArc",
-                    "MeshRemodelCreateWire",
-                    "MeshRemodelCreateTugBoat",
-                    "MeshRemodelRotateObject",
-                    "MeshRemodelMoveAxial",
-                    "MeshRemodelGoBackSelection",
-                    "MeshRemodelGroupCommandExtras",
-                    # "MeshRemodelMergeSketches",
-                    # "MeshRemodelValidateSketch",
-                    # "MeshRemodelPartCheckGeometry",
-                    # "MeshRemodelSubShapeBinder",
-                    # "MeshRemodelSettings",
-
-                    ] # A list of command names created in the line above
-        self.appendToolbar("MeshRemodel Commands",self.list) # leave settings, validate sketch and merge sketch off toolbar
-        self.appendMenu("Mesh&Remodel",self.list) # creates a new menu
-        #considered putting the menu inside the Edit menu, but decided against it
-        #self.appendMenu(["&Edit","MeshRemodel"],self.list) # appends a submenu to an existing menu
-
-
-    def Activated(self):
-        "This function is executed when the workbench is activated"
-        import requests
-        import xml.etree.ElementTree as ET
-
-        def get_remote_version(user, repo, branch='master'):
-            # GitHub raw URL for package.xml
-            url = f"https://raw.githubusercontent.com/{user}/{repo}/{branch}/package.xml"
-            try:
-                response = requests.get(url)
-                if response.status_code == 200:
-                    # Parse the XML content
-                    xml_content = response.content
-                    tree = ET.ElementTree(ET.fromstring(xml_content))
-                    root = tree.getroot()
-                    # Find the version element and return its text
-                    version = root.find("version").text
-                    return version
-                else:
-                    print(f"Failed to fetch package.xml: {response.status_code}")
-                    return None
-            except Exception as e:
-                print(f"Error fetching or parsing package.xml: {e}")
-                return None
-
-        def check_for_update(current_version, user, repo, branch, callback):
-            latest_version = get_remote_version(user, repo, branch)
-            if latest_version and latest_version > current_version:
-                callback(latest_version)
-
-        # Example usage
-        def update_callback(latest_version):
-            Console.PrintWarning(f"MeshRemodel {latest_version} is now available in the Addon Manager.\n")
-
-        from .MeshRemodelCmd import __version__ as current_version
-        user = "mwganson"
-        repo = "MeshRemodel"
-        branch = "master"
-
-        # Check for updates
-        pg = ParamGet("User parameter:/Plugins/MeshRemodel")
-        checkUpdates = pg.GetBool("CheckForUpdates", True)
-        #print(f"checkUpdates = {checkUpdates}")
-        if checkUpdates:
-            check_for_update(current_version, user, repo, branch, update_callback)
-
-        return
-
-    def Deactivated(self):
-        "This function is executed when the workbench is deactivated"
-
-        #FreeCAD will hide our menu and toolbar upon exiting the wb, so we setup a singleshot
-        #to unhide them once FreeCAD is finished, 2 seconds later
-        from .Qt import Core as QtCore
-        QtCore.QTimer.singleShot(2000, self.showMenu)
-        return
-
-    def askAboutToolbar(self):
-
-        from .Qt import Widgets as QtWidgets
-
-        parent=Gui.getMainWindow()
-        title = "Keep toolbar active?"
-        message = \
-"""
-In MeshRemodel we can keep the toolbar active
-when leaving the workbench.  Would you like to
-enable this 'keep the toolbar active' feature?
-
-If you choose Enable always or Disable always
-you can re-enable this popup question in the
-MeshRemodel settings dialog.
-"""
-        msgBox = QtWidgets.QMessageBox(parent)
-        msgBox.setWindowTitle(title)
-        msgBox.setText(message)
-
-        enableAlwaysButton = msgBox.addButton("Enable always", QtWidgets.QMessageBox.ButtonRole.AcceptRole)
-        enableOnceButton = msgBox.addButton("Enable this time", QtWidgets.QMessageBox.ButtonRole.RejectRole)
-        disableAlwaysButton = msgBox.addButton("Disable always", QtWidgets.QMessageBox.ButtonRole.AcceptRole)
-        disableOnceButton = msgBox.addButton("Disable this time", QtWidgets.QMessageBox.ButtonRole.AcceptRole)
-        msgBox.setDefaultButton(enableOnceButton)
-
-        msgBox.exec_()
-
-        if msgBox.clickedButton() == enableAlwaysButton:
-            clickedButton = "Enable always"
-        elif msgBox.clickedButton() == enableOnceButton:
-            clickedButton = "Enable once"
-        elif msgBox.clickedButton() == disableAlwaysButton:
-            clickedButton = "Disable always"
-        elif msgBox.clickedButton() == disableOnceButton:
-            clickedButton = "Disable once"
-        else:
-            clickedButton = None
-
-        return clickedButton
-
-    def showMenu(self):
-        from .Qt import Widgets as QtWidgets
-        window = QtWidgets.QApplication.activeWindow()
-        #freecad hides wb toolbars on leaving wb, we unhide ours here to keep it around
-        #if the user has it set in parameters to do so
-        pg = ParamGet("User parameter:Plugins/MeshRemodel")
-        ask = pg.GetBool("AskToolbar", True)
-        keep = pg.GetBool('KeepToolbar',True)
-        if ask:
-            response = self.askAboutToolbar()
-            if response == "Disable once":
-                ask = True
-                keep = False
-            elif response == "Enable always":
-                ask = False
-                keep = True
-            elif response == "Enable once":
-                ask = True
-                keep = True
-            elif response == "Disable always":
-                ask = False
-                keep = False
-            pg.SetBool("KeepToolbar", keep)
-            pg.SetBool("AskToolbar", ask)
-
-        if not keep:
-            return
-        tb = window.findChildren(QtWidgets.QToolBar) if window else []
-        for bar in tb:
-            if "MeshRemodel Commands" in bar.objectName():
-                bar.setVisible(True)
-
-    def ContextMenu(self, recipient):
-        "This is executed whenever the user right-clicks on screen"
-        # "recipient" will be either "view" or "tree"
-        self.appendContextMenu("MeshRemodel",self.list) # add commands to the context menu
-
-    def GetClassName(self):
-        # this function is mandatory if this is a full python workbench
-        return "Gui::PythonWorkbench"
-wb = MeshRemodelWorkbench()
-Gui.addWorkbench(wb)
-
-
-
-
+import freecad.Mesh_Remodel.MeshRemodelCmd
+import freecad.Mesh_Remodel.Workbench
